requirements.txt
pytest
black
pylint
PyQt6
pillow
pyinstaller
